requirements.txt
aiohttp
mode-streaming
celery
kombu